industry: refactor "update" to avoid timeout
nginx gives us 30 seconds to reply to the request, and bpos_update takes
too long. Part of that is processing time and part of it is building the
response page. We can get "some" output by limiting our processing time
to, i.e., 15 seconds and hope the page builds in the next 15 (it does).
Now we only see about 50 BP's before we hit our timeout (don't even get
through the A's in the list) but at least it doesn't error out. Really
this work needs to be done offline and the results added to bpos_view.

I also reordered the loops so we only iterate BP's we care about. This
ends up seeming slower (I don't process as many BP's), maybe because
django's filter() is better than my manual filter? Or maybe just because
I moved the bigger iteration (the API BP list) to the inside so I get
fewer outer loops and it just seems slower, but is actually the same if
I could run to completion.

The output shows that we have a lot of BP's to correct after Crius,
however, until we also fix the algorithms that consume these numbers as
started in 33f7eaef5ee1fa66c315595ca8de2aeaf3267d39 in dev/scot/crius,
keeping the old numbers to work with the old algoritms is probably
sadly better.